Full refactor of backend before main pull
