+++ backend/core/websocket.py
@@ -1,0 +1,28 @@
+import json
+from typing import List
+from fastapi import WebSocket
+
+class ConnectionManager:
+    def __init__(self) -> None:
+        self.active: List[WebSocket] = []
+
+    async def connect(self, ws: WebSocket) -> None:
+        await ws.accept()
+        self.active.append(ws)
+
+    def disconnect(self, ws: WebSocket) -> None:
+        if ws in self.active:
+            self.active.remove(ws)
+
+    async def broadcast(self, payload: dict) -> None:
+        msg = json.dumps(payload)
+        dead = []
+        for ws in self.active:
+            try:
+                await ws.send_text(msg)
+            except Exception:
+                dead.append(ws)
+        for ws in dead:
+            self.disconnect(ws)
+
+manager = ConnectionManager()
