+++ backend/utils/time.py
@@ -1,0 +1,4 @@
+from datetime import datetime, timezone
+
+def now_iso() -> str:
+    return datetime.now(timezone.utc).isoformat()
